+++ packages/datasource-customizer/src/decorators/computed/types.ts
@@ -17,3 +17,10 @@ export interface ComputedDefinition<
     context: CollectionCustomizationContext<S, N>,
   ): Promise<unknown[]> | unknown[];
 }
+
+export interface DeprecatedComputedDefinition<
+  S extends TSchema = TSchema,
+  N extends TCollectionName<S> = TCollectionName<S>,
+> extends Omit<ComputedDefinition<S, N>, 'columnType'> {
+  readonly columnType: 'Timeonly';
+}