+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,20 @@
+## @forestadmin/agent [1.29.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.29.1...@forestadmin/agent@1.29.2) (2023-09-20)
+
+
+### Bug Fixes
+
+* deprecate timeonly custom field type and map it to time ([#827](https://github.com/ForestAdmin/agent-nodejs/issues/827)) ([0311f7d](https://github.com/ForestAdmin/agent-nodejs/commit/0311f7d8367abc4cd9aed7f73c8bbe09dc203821))
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.29.2
+* **@forestadmin/datasource-toolkit:** upgraded to 1.22.1
+* **@forestadmin/forestadmin-client:** upgraded to 1.18.1
+
 ## @forestadmin/agent [1.29.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.29.0...@forestadmin/agent@1.29.1) (2023-09-18)