+++ packages/datasource-dummy/CHANGELOG.md
@@ -1,3 +1,14 @@
+## @forestadmin/datasource-dummy [1.0.64](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.63...@forestadmin/datasource-dummy@1.0.64) (2023-09-20)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.29.2
+* **@forestadmin/datasource-toolkit:** upgraded to 1.22.1
+
 ## @forestadmin/datasource-dummy [1.0.63](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.62...@forestadmin/datasource-dummy@1.0.63) (2023-09-18)