+++ packages/plugin-aws-s3/CHANGELOG.md
@@ -1,3 +1,14 @@
+## @forestadmin/plugin-aws-s3 [1.3.28](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.27...@forestadmin/plugin-aws-s3@1.3.28) (2023-09-20)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.29.2
+* **@forestadmin/datasource-toolkit:** upgraded to 1.22.1
+
 ## @forestadmin/plugin-aws-s3 [1.3.27](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.26...@forestadmin/plugin-aws-s3@1.3.27) (2023-09-18)